# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puryaideu Api",
        "description": "Implementation of Swagger with in Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/location_riders": {
            "post": {
                "tags": [
                    "Riders"
                ],
                "summary": "Location",
                "operationId": "location",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rider_id": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Riders location stored",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/rider_register": {
            "post": {
                "tags": [
                    "Riders"
                ],
                "summary": "Register Rider",
                "operationId": "register rider",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "profile_pic": {
                                        "type": "string",
                                        "format": "file"
                                    },
                                    "licence": {
                                        "type": "file"
                                    },
                                    "licence_number": {
                                        "type": "string"
                                    },
                                    "experience": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Deepika",
                                    "email": "deepik@gmail.com",
                                    "phone": "+977 9812323132",
                                    "profile_pic": "2312.jpg",
                                    "license": "license.png",
                                    "license_number": "2323132",
                                    "experience": "2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Registration",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/specific_user": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::user_id",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User details displayed successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}